requirements.txt
streamlit
pandas
numpy
nltk
scikit-learn
beautifulsoup4
requests
python-docx
